requirements.txt
pandas
Jinja2
python-dotenv

for terminal:
(pip install -r requirements.txt)